Hotfix de campo horas en los proyectos

+++ isc.time.report.be/isc.time.report.be.domain/Entity/Projects/Project.cs
@@ -26,7 +26,7 @@ namespace isc.time.report.be.domain.Entity.Projects
         public DateTime? ActualStartDate { get; set; }
         public DateTime? ActualEndDate { get; set; }
         public decimal? Budget { get; set; }
-        public int Hours { get; set; }
+        public decimal Hours { get; set; }
         public List<EmployeeProject> EmployeeProject { get; set; }
     }
 }
